+++ libbeat/common/transport/tls.go
@@ -21,7 +21,6 @@ package transport
 
 import (
 	"crypto/tls"
-	"errors"
 	"fmt"
 	"net"
 	"sync"
@@ -30,10 +29,6 @@ import (
 	"github.com/elastic/beats/v7/libbeat/common/transport/tlscommon"
 	"github.com/elastic/beats/v7/libbeat/testing"
 )
-
-func TLSDialer(forward Dialer, config *tlscommon.TLSConfig, timeout time.Duration) Dialer {
-	return TestTLSDialer(testing.NullDriver, forward, config, timeout)
-}
 
 func TestTLSDialer(
 	d testing.Driver,
@@ -130,87 +125,4 @@ func TestTLSDialerH2(
 
 		return tlsDialWith(d, forward, network, address, timeout, tlsConfig, config)
 	}), nil
-}
-
-func tlsDialWith(
-	d testing.Driver,
-	dialer Dialer,
-	network, address string,
-	timeout time.Duration,
-	tlsConfig *tls.Config,
-	config *tlscommon.TLSConfig,
-) (net.Conn, error) {
-	socket, err := dialer.Dial(network, address)
-	if err != nil {
-		return nil, err
-	}
-
-	conn := tls.Client(socket, tlsConfig)
-
-	withTimeout := timeout > 0
-	if withTimeout {
-		if err := conn.SetDeadline(time.Now().Add(timeout)); err != nil {
-			d.Fatal("timeout", err)
-			_ = conn.Close()
-			return nil, err
-		}
-	}
-
-	if tlsConfig.InsecureSkipVerify {
-		d.Warn("security", "server's certificate chain verification is disabled")
-	} else {
-		d.Info("security", "server's certificate chain verification is enabled")
-	}
-
-	err = conn.Handshake()
-	d.Fatal("handshake", err)
-	if err != nil {
-		_ = conn.Close()
-		return nil, err
-	}
-
-	// remove timeout if handshake was subject to timeout:
-	if withTimeout {
-		conn.SetDeadline(time.Time{})
-	}
-
-	if err := postVerifyTLSConnection(d, conn, config); err != nil {
-		_ = conn.Close()
-		return nil, err
-	}
-
-	return conn, nil
-}
-
-func postVerifyTLSConnection(d testing.Driver, conn *tls.Conn, config *tlscommon.TLSConfig) error {
-	st := conn.ConnectionState()
-
-	if !st.HandshakeComplete {
-		err := errors.New("incomplete handshake")
-		d.Fatal("incomplete handshake", err)
-		return err
-	}
-
-	d.Info("TLS version", fmt.Sprintf("%v", tlscommon.TLSVersion(st.Version)))
-
-	// no more checks if no extra configs available
-	if config == nil {
-		return nil
-	}
-
-	versions := config.Versions
-	if versions == nil {
-		versions = tlscommon.TLSDefaultVersions
-	}
-	versionOK := false
-	for _, version := range versions {
-		versionOK = versionOK || st.Version == uint16(version)
-	}
-	if !versionOK {
-		err := fmt.Errorf("tls version %v not configured", tlscommon.TLSVersion(st.Version))
-		d.Fatal("TLS version", err)
-		return err
-	}
-
-	return nil
 }
