+++ libbeat/common/transport/tlscommon/config.go
@@ -104,8 +104,3 @@ func (c *Config) Validate() error {
 
 	return c.Certificate.Validate()
 }
-
-// IsEnabled returns true if the `enable` field is set to true in the yaml.
-func (c *Config) IsEnabled() bool {
-	return c != nil && (c.Enabled == nil || *c.Enabled)
-}
