+++ libbeat/common/transport/tlscommon/server_config_legacy.go
@@ -23,8 +23,6 @@ import (
 	"crypto/tls"
 
 	"github.com/joeshaw/multierror"
-
-	"github.com/elastic/beats/v7/libbeat/common"
 )
 
 // ServerConfig defines the user configurable tls options for any TCP based service.
@@ -91,42 +89,4 @@ func LoadTLSServerConfig(config *ServerConfig) (*TLSConfig, error) {
 		CurvePreferences: curves,
 		ClientAuth:       tls.ClientAuthType(config.ClientAuth),
 	}, nil
-}
-
-// Unpack unpacks the TLS Server configuration.
-func (c *ServerConfig) Unpack(cfg common.Config) error {
-	const clientAuthKey = "client_authentication"
-	const ca = "certificate_authorities"
-
-	// When we have explicitely defined the `certificate_authorities` in the configuration we default
-	// to `required` for the `client_authentication`, when CA is not defined we should set to `none`.
-	if cfg.HasField(ca) && !cfg.HasField(clientAuthKey) {
-		cfg.SetString(clientAuthKey, -1, "required")
-	}
-	type serverCfg ServerConfig
-	var sCfg serverCfg
-	if err := cfg.Unpack(&sCfg); err != nil {
-		return err
-	}
-	*c = ServerConfig(sCfg)
-	return nil
-}
-
-// Validate values the TLSConfig struct making sure certificate sure we have both a certificate and
-// a key.
-func (c *ServerConfig) Validate() error {
-	if c.IsEnabled() {
-		// c.Certificate.Validate() ensures that both a certificate and key
-		// are specified, or neither are specified. For server-side TLS we
-		// require both to be specified.
-		if c.Certificate.Certificate == "" {
-			return ErrCertificateUnspecified
-		}
-	}
-	return c.Certificate.Validate()
-}
-
-// IsEnabled returns true if the `enable` field is set to true in the yaml.
-func (c *ServerConfig) IsEnabled() bool {
-	return c != nil && (c.Enabled == nil || *c.Enabled)
 }
